pyde-doc-rope: Check the preceding symbol if currently no doc found.
When in the arguments of a function, the old version simply didn't
find documentation for the function. This new version will move
backwards in the hope of finding a symbol to show documentation
for there.

This also handles the corner case where `rope-get-doc` returns "",
which is now treated as "no documentation" and not shown.

Fixes #7.