elpy-flymake-python-init: Ignore remote files.
Flymake has serious problems with remote files. The code breaks,
and most importantly, tramp will try to run the check command
remotely, which might not even work at all. This change means that
remote files are not checked at all with flymake, but it removes
a lot of errors and annoyances.