rope get_completions, get_calltip: Handle ModuleSyntaxError
When there are too many syntax errors in a file, Rope can not fix
them all. In that case, it throws a very verbose error.

For get_completions, we catch that error and make it a bit cleaner.
For get_calltip, we catch the error and simply ignore it, as not
to annoy the user too much while typing.