elpy-initialize-variables: New function.
This reworks the way elpy overwrites the variables of other modes.
The old method tried to be too smart, but relied on elpy being
loaded first to actually work. The new method might overwrite
user configurations if `elpy-enable` is run after those variables
are set, but that should be more accessible and understandable
than the old behavior.

Fixes #35