forward-sexp-function: Set to nil in python-mode.
The default behavior of python-mode is very strange and inconsistent,
as it considers the last line of a method to be "the end" of the method
and moving backwards a sexp there will not move backwards over a symbol,
but backwards to the beginning of the method. The maintainer considers
this a feature and thinks setting `forward-sexp-function` to nil is
the way to fix it for users.

Elpy does this now, but it required a new function to be run in the
python hook.

See http://debbugs.gnu.org/db/13/13642.html for the report and
discussion.