elpy-initialize-variables: Bind RET as no-op in ac-completion-map
When using auto-complete as a completion library for "search as you
type" style completion, using RET usually means the user is happy
with the current completion and just wants a newline, much like
SPC means the user is happy with the current contents and wants
a space inserted. The default behavior of ac uses RET to confirm
the current completion, which can be very annoying.

This commit changes this default behavior.