elpy-refactor-window-configuration: New variable.
In a refactoring buffer, store the window configuration from when
the refactoring action was started, and restore it when the
operation is committed. This prevents window configurations from
being lost after a refactoring action.

This commit also records point in buffers being changed to
leave it at least somewhat closeby if the content changes.

Fixes #45.