elpy-shell-send-region-or-buffer: Pass prefix arg to python-shell-send-buffer
The new python.el's python-shell-send-buffer function is smart and
replaces if __name__ == '__main__' lines with something that does
not match, to avoid executing code that's meant to run only in a
script. You can force that to be executed by passing a prefix arg.

This is a nice idea, so elpy now does the same, passing the prefix
arg to the python.el function.

Fixes #69.