elpy-rpc-project-specific: Remove the 'ask option.
This was very annoying, as the RPC process is not even started
when the buffer is opened. Suddenly, a prompt would show up. So
I changed my mind on the usefulness of this option. People who want
the 'ask behavior can set the value to nil and simply run
`M-x elpy-rpc-open-project-specific` when they want a project-specific
Elpy.

`elpy-rpc-pool-open` adjusted accordingly.