Inferior Python Shell: Do not prompt for start and switch when sending.
With this change, elpy will start the inferior Python process without
prompting for a command and whether to make it dedicated. If you need
this functionality, you can use M-x run-python directly.

In elpy, C-c C-c will now send the current region or buffer to the
global inferior process and switch to it, starting a new process
if necessary. C-c C-z will just switch to the inferior process,
creating it if necessary.

Fixes #68
Fixes #85