flake8 support: Improve flake8 detection and warning display.
First, we set python-check-command to flake8 both when flake8 is
found and when python-check-command is not set to an already
existing command. This is useful for when we have flake8 installed
only in virtualenvs, so not when elpy starts for the first time.

We also use the new variable name for flymake warnings that Emacs
24.4 introduced.

This might or might not help with #57.