Be more explicit about version numbers.
This commit provides M-x elpy-version, which will show the version
of the Elpy package both in Emacs Lisp as well as the RPC backend
if we are in an Elpy buffer.

This implements #138.

We also now claim to be the version "devel" in the git repository,
making it easier to detect mismatches between frontend and backend
versions. This should make it easier to detect problems due to
a frontend version from a different source than the backend.

This, together with MELPA hopefully sticking to the release tag,
should implement #139.