Send refactor RPC request args as a list.
The 'params' of the request was being set to the first arg. Since this was
a string (the project root path, specifically), the arguments to the
server methods were being set to the list of characters in the path. This
ultimatedly caused Emacs error messages like "error in process filter:
Invalid function 12345" when trying to refactor.