Use evm in .travis.yml
The ppa for emacs-snapshot is not being updated anymore. Also, it
has a broken version of Emacs in it that does not work with cask.
And finally, it only had one version of Emacs 24.x anyhow.

So now we use evm to provision different versions of Emacs for
testing.

Snapshot versions of Emacs will have to be tested outside of
Travis.