Start elpy-rpc process in the current buffer's context
If a user made process-environment buffer local and added environment
variables of his own to support his use-case, those changes would be
lost to the elpy-rpc process if it was started in the context of a
new buffer (like before).

Possible scenario: A dedicated virtualenv for all emacs related dev
packages (elpy, PyLint, ...) which is added via PYTHONPATH to each
opened Python buffer. In that buffer, exec-path and process-environment
are buffer local to isolate those changes and thus make it possible to
work on projects with different Python versions and/or virtualenvs.

NB: start-process always starts the process with the current buffer's
process-environment, even if a new destination buffer is given as an
argument.