Ensure that auto-complete documentation is not the empty string.
Auto-complete uses popup.el to display documentation, which apparently
can not deal with empty strings in this case. An empty string will
cause an error/loop, both getting Emacs to consume CPU as well as
not removing any empty lines at the end of the buffer added by popup.el
to display the help popup.

Elpy now ensures that this never returns the empty string.

Thanks to Matthias Dahl for his work in finding this bug (#183).

Fixes #182 and possibly also #154.