Use pythonw on Windows systems
Apparently, the normal python command starts a DOS window, which
can cause hangs and freezes when used for Elpy's RPC process. On
Windows systems, Elpy will now use the pythonw command by default.

See #199 for a discussion of this issue. Thanks to users adamc55
and tiktok7 for trying to figure out the problem and finding a
solution.