Rework elpy configuration.
Elpy will now not pop up an installation screen on any configuration
problem anymore, but simply signal an error and suggest the user
to run M-x elpy-config. This command will open a buffer with some
information about the Elpy configuration and customization options
for Elpy.

To make this more useful, the Elpy customization options have been
reworked.

This change removes elpy-show-installation-instructions, which is
now useless.

This implements most of #146.