Do not select the shell buffer after sending the current buffer.
This is not always what users want, as they can include side effects
in their module, and want to just see the results of evaluating the
code, not actually interact with it. Because of this,
elpy-shell-send-region-or-buffer now displays the inferior shell
buffer, but does not select it.

Users who want to switch to the buffer after sending some code can
just use C-c C-c followed by C-c C-z.