Error handling rework.
When Elpy encounters a problem it can't handle, especially in the
backend, it will now pop up a buffer with configuration options
and backtrace information, as well as a link to the bug tracker.

The RPC server now also provides a way of throwing a warning
exception only, resulting in a simple message without any actual
errors, which should result in a much more smooth interaction
experience.