Rework elpy-doc.
The function will now display the documentation for the symbol at point.
If it can't find the documentation, it'll ask pydoc. If that does not
work, either, or if the user specified a prefix arg, it'll ask the
user for a symbol to show pydoc on, providing tab completion for
possible symbols.

This also removes the websearch functionality, which was limited
at best to begin with.