Distinguish between project and library root.
Elpy now has two different concepts for "root directories". The project
root contains all files of a project, while the library root is the
base directory from which the current file could be imported. For many
projects, these are the same, but that's not always the case.

The project root interface now has a more elaborate way of being
configured, with a number of different options.

This also fixes #235.