Unify RPC to use JSON-RPC error messages.
There was a lot of special cases for the error handling when it
comes to tracebacks and other additional information, all the while
not utilizing e.g. the code argument for JSON-RPC errors. This
rework utilizes the code argument fully to specify what type of
error this is (warning, error, error with traceback), and uses the
official data element to pass around extra information like
tracebacks.