Rework RPC interface for a simpler initialization.
The native backend has been removed. The only functionality it
added was rlcomplete-based completion, which was severely limited.

The initialization process of backends has been simplified. This
will allow easier use of state for future extensions.

The simplification made elpy-set-backend obsolete. If you want to
change the backend, set elpy-rpc-backend and restart the RPC
processes.