Extract documentation from completions, add locations.
Elpy now supports both the documentation and the location display
feature of Company. You can use C-d to see documentation for a
completion and C-w to see the source in context.

This change also removes docstrings from the default completion
response, improving performance in the common case, and finishing
the transition from auto-complete to company-mode.