Make Rope work with unusually large directories.
Rope will happily recurse into all subdirectories of the project
root. If a user opens a file in their home directory, this might be
quite a lot of work, making Rope mostly unusable in those
circumstances. Elpy now patches Rope so it will not recurse into
directories that do not have __init__.py files, speeding up
the process considerably. As Elpy will pass the library root as
the project root, this will do the Right Thing for elpy.