Rope will not error when the project root is missing.
It was possible, especially in tramp buffers, for Elpy to initialize
Rope with an inexisting project root. This would make Rope throw
an error. We now initialize Rope with the empty string as the project
root.

I'm not sure if the empty string is the best option, but Rope does
not like None, and at this point, we have no better option, really.

Fixes #319.