Check that we found a mode lighter entry before setting it.
It is possible that a mode lighter is not in the mode line, for
example when the mode name changes upstream or the user removed
the mode from the mode line themselves. Guard against this so we
do not throw an error.

Fixes the error described at the end of #477.