Remove iblock movement.
This was difficult to get right and caused a lot of confusion in
edge cases, so just ask the user to mark the block they want to use.
Much clearer.

If the original behavior is still desired, we could add some code
that marks the current block instead.

Fixes #367, fixes #426, fixes #443