Complete callables optionally with parentheses
Resolve issue https://github.com/jorgenschaefer/elpy/issues/643, since
point is left between the parentheses.

Provide a single post completion function to sensibly (aims to be
correct in 90 % of the cases) complete callables with parentheses.

Allow the user either to disable the completion by configuring an empty
completion function or to provide his own post completion function to
adapt to his Emacs and Python environments, and to his coding style.