Disable error display by default
Jedi has a ton of bugs by now and no new release is upcoming,
apparently. The load of error reports is nice if someone would fix
them, but as is, all Elpy could do is to add ignores anyhow. So we
disable displaying of errors by default, which amounts to the
same, and wait for a new Jedi release.