Add basic Django support
You can now parse through `manage.py` (If it is found, otherwise it
falls back to `django-admin.py`) for commands to run and emacs will
provide completion assistance.

Another function is to automatically start the development server
command, `runserver`, with default variables that are set to
`127.0.0.1`(for ip address) and `8000` (for port). Each Django project
has the ability of starting their own dev server.

One question I have regarding this is if we should also allow changing
the command for starting the development server, `runserver`? I know
other people like `runserver_plus` from the `django-extensions` module.

Commit also introduces a new dependency, `s`, for string
manipulation. Currently, it is only used on one line for trimming
whitespace. Should the commit introduce this new dep or simply create
a function that does this?