Fix bug in elpy-shell-kill-all where it does not find the *Python* buffer(s)
This is a fix for the problem reported in
https://github.com/jorgenschaefer/elpy/issues/1147#issue-238358628

This replaces the original regular-expression for one that does not
mandate that the buffer name contains a square-bracketed
expression. It is unknown why it needed to match the square-bracketed
expression in the first place, since the only buffer names I have seen
in practice being created by elpy are "*Python*". Perhaps in the past
there were allowed to be multiple interactive Python buffers?
Regardless, make it match either with or without the square-bracketed
form of buffer names.