Fix typo in elpy-shell-add-to-shell-history docstring & enhance flow
Post-edits, this docstring makes it sound like this toggle enables a
kill-ring or clipboard history for the Python shell, which sounds like
it's always a good thing.  That said, it's disabled by default.  I
believe it would be useful to let the user know why it's disabled by
default.

Also, I think it would be useful to have a follow up commit that
explains how it "affects the following functions:…".