Fix #1823: support newer eldoc protocol
This UNTESTED commit should hint at how to fix the issue without
resorting to `eldoc-docstring-format-sym-doc', which is gone in Emacs
28 (ElDoc 1.1.0).

Beyond the ideas delineated in this commit, the new ElDoc protocol
would also easily obviate the need to have "flymake-specific" ElDoc
code in ElPy (this is now default in stock Flymake), as well as being
forced to choose between different types of ElPy docstrings.

* elpy.el (elpy-module-eldoc): Add to eldoc-documentation-functions.
(elpy-eldoc-documentation): Rework to support
eldoc-documentation-functions.