Separate super and sub classes' snippets
The way the snippet was set (one only snippet for both
super and sub classes) prompted the creation of super classes
of the kind:

class ClassName(object):
      '''Doc'''
      def __init__(self):
      	  super().__init__()
	  ...

calling the method super() in a subclass of 'object'
rises the 'useless-super-delegation' warning from
pylintrc.
This can be avoided by creating a separated snippet
for subclasses, in which the method super() has an
usefull effect.

Signed-off-by: Gustavo Gurgel <gukgurgel@gmail.com>