Remove Patreon link
Dear patrons,

It has come to a point where I have to face reality and be open and
transparent to all of you. I have stopped using Emacs and with that,
also stopped supporting the various projects I work on. Most if not
all of you are on Patreon for me because of my Emacs projects, so
it's only fair to let you know that your contributions might be more
useful when directed to other creators.

Thank you all for your absolutely stunning support. I would not have
thought it possible to make a small amount of money through work I
loved doing. I am truly grateful for your generosity.

Jorgen