Fix #1867 - allow MacOS users to choose pty or pipe for Python shell
This change creates a custom variable that allows MacOS users to decide
whether to use a pipe or a pseudo-terminal (pty) to connect to the Python
shell when `elpy-shell-get-or-create-process` is called. Before this commit,
all Mac users are forced to use a pipe if the Python shell is created using
this function, even though this is not necessary for all users.

Some MacOS users experience issues connecting to the Python shell using a
pty. One of major problems is the inability to send greater than 1024
characters to the shell at one time. These issues are caused by versions of
Python that use the libedit version of Readline rather than the GNU
version. By default, MacOS does not have GNU Readline installed, so it instead
uses the libedit version for the versions of Python built into the OS. This
may also be the case for other versions installed on a Mac if GNU Readline was
not installed first. Using a pipe instead a pty alleviates all of these issues
at the expense of creating a few smaller issues (e.g. sending an EOF signal
into pdb causes the entire shell to exit).

However, many MacOS users use versions of Python that are built with GNU
Readline which means that the pty will work properly for them. It is
unnecessary for them to connect via a pipe. Additionally, users may
run into confusing behavior where the Python shell created via `run-python`
(which uses a pty unless `process-connection-type` is set to nil) acts
differently than one that is created via an Elpy command.

Since a pipe connection to the Python shell works for users with and without
versions of Python installed with GNU readline, this continues to be a sane
default for Mac users. Creating this custom variable provides an easy, if
somewhat difficult to discover, way to allow Mac users to decide which process
type they want to use.

In the future, it may be best to ask Mac users on their first usage of Elpy
whether to use a pipe or pty. Alternatively, this could be automatically
detected and a message could be sent to the Python interpreter to suggest that
they change this variable's value.

In addition to #1867, this change is also a fix for #1907 and #1948.

Other related issues and commits include: #887, #1550, #1838, and #1671.